feat(sandbox): promote sandbox from --beta flag to first-class cloud
Running an agent in a local Docker container is now a cloud target:
`spawn <agent> sandbox`, instead of `spawn <agent> local --beta sandbox`.

- Extract local orchestration into local/run.ts (runLocalAgent), shared
  by the local and sandbox clouds; add sandbox/main.ts entry point.
- Add 9 sh/sandbox/*.sh shims + README for the Docker-capable agents.
- Register the sandbox cloud and matrix entries in manifest.json.
- Remove sandbox from the --beta gate and the fast_provision experiment.
- Bump CLI to 1.1.0; update tests for the new wiring.
