+++ manifest.json
@@ -392,6 +392,18 @@
       "interactive_method": "exec",
       "notes": "No cloud provisioning needed. Installs agents and injects OpenRouter credentials locally. Useful for local development and testing."
     },
+    "sandbox": {
+      "name": "Local Sandbox",
+      "price": "Free",
+      "description": "Your computer, in a throwaway Docker container — no account or payment needed",
+      "url": "https://github.com/OpenRouterTeam/spawn",
+      "type": "local",
+      "auth": "none",
+      "provision_method": "docker run (local container)",
+      "exec_method": "docker exec",
+      "interactive_method": "docker exec -it",
+      "notes": "Runs the agent inside an isolated Docker container on your machine. Docker is auto-installed if missing, and the container is removed on exit. Same setup as the local cloud, but the agent cannot touch your host filesystem."
+    },
     "hetzner": {
       "name": "Hetzner Cloud",
       "price": "~€3/mo",
@@ -498,6 +510,16 @@
     "local/codex": "implemented",
     "local/opencode": "implemented",
     "local/kilocode": "implemented",
+    "sandbox/claude": "implemented",
+    "sandbox/openclaw": "implemented",
+    "sandbox/codex": "implemented",
+    "sandbox/opencode": "implemented",
+    "sandbox/kilocode": "implemented",
+    "sandbox/hermes": "implemented",
+    "sandbox/junie": "implemented",
+    "sandbox/pi": "implemented",
+    "sandbox/cursor": "implemented",
+    "sandbox/t3code": "missing",
     "hetzner/claude": "implemented",
     "hetzner/openclaw": "implemented",
     "hetzner/codex": "implemented",
