+++ packages/cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@openrouter/spawn",
-  "version": "1.0.44",
+  "version": "1.1.0",
   "type": "module",
   "bin": {
     "spawn": "cli.js"
