+++ packages/cli/src/local/main.ts
@@ -1,27 +1,15 @@
 #!/usr/bin/env bun
 
-// local/main.ts — Orchestrator: deploys an agent on the local machine
+// local/main.ts — Orchestrator: deploys an agent on the local machine.
+//
+// For the isolated Docker-container variant, see sandbox/main.ts — both share
+// the orchestration in local/run.ts.
 
-import type { CloudOrchestrator } from "../shared/orchestrate.js";
-
-import * as p from "@clack/prompts";
 import { getErrorMessage } from "@openrouter/spawn-shared";
 import pkg from "../../package.json" with { type: "json" };
-import { createCloudAgents } from "../shared/agent-setup.js";
-import { makeDockerRunner, runOrchestration } from "../shared/orchestrate.js";
 import { initTelemetry } from "../shared/telemetry.js";
-import { logWarn } from "../shared/ui.js";
-import { agents, resolveAgent } from "./agents.js";
-import {
-  cleanupContainer,
-  dockerInteractiveSession,
-  downloadFile,
-  ensureDocker,
-  interactiveSession,
-  pullAndStartContainer,
-  runLocal,
-  uploadFile,
-} from "./local.js";
+import { agents } from "./agents.js";
+import { runLocalAgent } from "./run.js";
 
 async function main() {
   const agentName = process.argv[2];
@@ -31,92 +19,7 @@ async function main() {
     process.exit(1);
   }
 
-  // Check if --beta sandbox is active
-  const betaFeatures = (process.env.SPAWN_BETA ?? "").split(",");
-  const useSandbox = betaFeatures.includes("sandbox");
-
-  const baseRunner = {
-    runServer: runLocal,
-    uploadFile: async (l: string, r: string) => uploadFile(l, r),
-    downloadFile: async (r: string, l: string) => downloadFile(r, l),
-  };
-
-  // When sandboxed, recreate agents with the Docker-wrapped runner so that
-  // agent.configure() / agent.install() closures execute inside the container
-  // instead of writing config files directly to the host filesystem.
-  const agent = useSandbox
-    ? createCloudAgents(makeDockerRunner(baseRunner)).resolveAgent(agentName)
-    : resolveAgent(agentName);
-
-  // If sandboxed, ensure Docker is installed (auto-install if missing)
-  if (useSandbox) {
-    await ensureDocker();
-  }
-
-  // Warn about security implications of installing OpenClaw locally
-  // (skip warning in sandbox mode — the container provides isolation)
-  if (agentName === "openclaw" && !useSandbox && process.env.SPAWN_NON_INTERACTIVE !== "1") {
-    process.stderr.write("\n");
-    logWarn("⚠  Local installation warning");
-    logWarn(`   This will install ${agent.name} directly on your machine.`);
-    logWarn("   The agent will have full access to your filesystem, shell, and network.");
-    logWarn("   For isolation, consider running on a cloud VM instead.\n");
-
-    const confirmed = await p.confirm({
-      message: "Continue with local installation?",
-      initialValue: true,
-    });
-
-    if (p.isCancel(confirmed) || !confirmed) {
-      p.log.info("Installation cancelled.");
-      process.exit(0);
-    }
-  }
-
-  const cloud: CloudOrchestrator = {
-    cloudName: "local",
-    cloudLabel: useSandbox ? "local (sandboxed)" : "local",
-    skipAgentInstall: false,
-    runner: useSandbox ? makeDockerRunner(baseRunner) : baseRunner,
-    async authenticate() {},
-    async promptSize() {},
-    async createServer(_name: string) {
-      return {
-        ip: "localhost",
-        user: process.env.USER || "local",
-        cloud: "local",
-      };
-    },
-    async getServerName() {
-      const result = Bun.spawnSync(
-        [
-          "hostname",
-        ],
-        {
-          stdio: [
-            "ignore",
-            "pipe",
-            "ignore",
-          ],
-        },
-      );
-      return new TextDecoder().decode(result.stdout).trim() || "local";
-    },
-    async waitForReady() {
-      if (useSandbox) {
-        await pullAndStartContainer(agentName);
-        cloud.skipAgentInstall = true;
-      }
-    },
-    interactiveSession: useSandbox ? dockerInteractiveSession : interactiveSession,
-  };
-
-  // Clean up sandbox container on exit
-  if (useSandbox) {
-    process.on("exit", cleanupContainer);
-  }
-
-  await runOrchestration(cloud, agent, agentName);
+  await runLocalAgent(agentName, false);
 }
 
 initTelemetry(pkg.version);
