+++ packages/cli/src/shared/feature-flags.ts
@@ -215,7 +215,6 @@ export function expandFastProvisionVariant(variant: string): readonly string[] {
     return [
       "images",
       "docker",
-      "sandbox",
     ];
   }
   return [];
